requirements.txt
discord.py==2.3.2
vk-api==11.9.9